[추가] 보스 만들기

+++ My project (1)/Assets/02_script/DamageObject.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class DamageObject : MonoBehaviour
 {
-    public int Value_int;
+    public float Value_int;
 
     void OnCollisionEnter2D(Collision2D collision)
     {
